+++ pyviews/core/rendering.py
@@ -1,3 +1,5 @@
+"""dependency methods for rendering"""
+
 from injectool import dependency, DependencyError
 
 from .xml import XmlNode
@@ -5,12 +7,12 @@ from .node import Node
 
 
 @dependency()
-def create_node(xml_node: XmlNode, **init_args) -> Node:
+def create_node(xml_node: XmlNode, context: dict) -> Node:
     """Creates node from xml node using namespace as module and tag name as class name"""
-    raise DependencyError(xml_node, init_args)
+    raise DependencyError(xml_node, context)
 
 
 @dependency()
-def render(xml_node: XmlNode, **render_args) -> Node:
+def render(xml_node: XmlNode, context: dict) -> Node:
     """Renders node from xml node"""
-    raise DependencyError(xml_node, render_args)
+    raise DependencyError(xml_node, context)