+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Expression resolved using code as resolve parameter
 - Migrated to injectool 1.1.1
+- **args are changed to RenderingContext for rendering steps
 
 ## 2.1.1